more changessss

+++ client/scripts/controllers/view_resource.js
@@ -82,13 +82,30 @@ angular.module('resourceTracker')
 		};
 	
 		$scope.deleteResource = function() {
+			if (!validateDeleteResource()) {
+				return;
+			}
 			var deleteQueryString = '/resource?resource_id=' + $scope.editingResource.resource_id;
-			console.log($scope.selectedResource);
 			$http.delete(deleteQueryString).then(function(response) {
 				showMessageAndReload("Successfully deleted resource!");
             }, function(error) {
 				$scope.addError(resourceService.alertMessages.resourceUpdatingFailed);
             });
+		};
+
+		var validateDeleteResource = function() {
+			var nowDate = new Date();
+			var conflictingReservation = false;
+			$scope.selectedResource.reservations.forEach(function(reservation) {
+				var endDate = new Date(reservation.end_time);
+				if (endDate >= nowDate) {
+					conflictingReservation = true;
+				}
+			});
+			var conflictingReservationMessage = "There is a current or future reservation for this resource. " +
+				"Do you want to delete the resource?";
+		    return !conflictingReservation ||
+		    	(conflictingReservation && confirm(conflictingReservationMessage));
 		};
 
 		$scope.updateResource = function() {
@@ -152,42 +169,11 @@ angular.module('resourceTracker')
 			$scope.clearError();
 			$http.get('/resource/all').then(function(response) {
 				$scope.allResources = response.data;
+				console.log("SCOPE ALL RESOURCES");
 				console.log($scope.allResources);
             }, function(error) {
 				console.log(error);
             });		
-		//	$http.get('/tag').then(function(response) {
-			//	console.log(response.data);
-            //}, function(error) {
-				//console.log(error);
-        //    });
-
-			var includedTags = ['tag1'];
-			var excludedTags = [];
-
-			var toSend = {includedTags: includedTags, excludedTags: excludedTags, start_time: 0, end_time: Number.MAX_VALUE};
-
-			//$http.post('/tag/filter', toSend).then(function(response) {	
-				//console.log(response.data);
-           // }, function(error) {
-			///	console.log(error);
-		///	});
-
-            var testDate = new Date();
-            console.log(testDate.valueOf());
-
-            var startDate = new Date(2016, 1, 7, 18, 53);
-            var endDate = new Date(2016, 1, 7, 18, 54);
-            var reservation = {
-                resource_id: 2,
-                start_time: startDate.valueOf(),
-                end_time: endDate.valueOf()
-            };
-          //  $http.put('/reservation', reservation).then(function(response) {
-			//	console.log(response);
-            //}, function(error) {
-				//console.log(error);
-            //});
 		};
 
 		getAllResources();
